Only walk fetched nodes when saving
When saving, we can only have altered the nodes which we've seen so
far.  Assuming the user has seen the old form as well as the new form,
we should have saved all relevant fields which the user may have
altered during this edit session, by walking over these fields.