Add support for knowing which components are children of ours
The new path allows us to discover which components are rendered as
children of ourselves.  This requires the child to register itself and
some computation to occur.  The necessary code has been abstracted in
Mixins.  You have to:
- In the parent:
  - Use the has-registered-children mixin
  - Add the attribute register=(action "registerChild" childModel) to the
  consuming child
  - Define registeredChildrenModels as an alias of the children array
  so we can proactively remove children.
- In the child:
  - Use the is-registered-child mixin

The parent will have its child components available in a computed
property array named childComponents.  This object currently contains
{ child, component } so you have access to the model and the component
in a structured way.  This api and/or is likely to change.

WIP

updated model to allow input validation

WIP bidirectional information passing test

Bidirectional syncing of keys for properties

The properties now sync bidirectionally.  This implementation groups
the keys and sends them to the parent components using computed
proprties.  This minimizes the chances for synchronization errors.

cleanup of debugging code

added inputState computed property for component

make sure there is an array to use .map for

Set wrapping is-required class on InputFields

The calculation of this field is not super generic.  It does work for
our current cases and it will likely work for most other cases.  This
styling may be updated in future versions in order to accomodate
dynamic styling based on fields configured in the database.

cleanup linting errors