Add support for knowing which components are children of ours
The new path allows us to discover which components are rendered as
children of ourselves.  This requires the child to register itself and
some computation to occur.  The necessary code has been abstracted in
Mixins.  You have to:
- In the parent:
  - Use the has-registered-children mixin
  - Add the attribute register=(action "registerChild" childModel) to the
  consuming child
  - Define registeredChildrenModels as an alias of the children array
  so we can proactively remove children.
- In the child:
  - Use the is-registered-child mixin

The parent will have its child components available in a computed
property array named childComponents.  This object currently contains
{ child, component } so you have access to the model and the component
in a structured way.  This api and/or is likely to change.